Fix: Next burn issue
